a first successful split of the code into separate scripts
